+++ iam.tf
@@ -1,13 +1,17 @@
 locals {
+  # filter all objects that define a single role
   iam_role = [for iam in var.iam : iam if can(iam.role)]
 
+  # filter all objects that define multiple roles and expand them to single roles
   iam_roles = flatten([for iam in var.iam :
     [for role in iam.roles : merge(iam, { role = role })] if can(iam.roles)
   ])
 
-  iam_map = { for iam in concat(local.iam_role, local.iam_roles) : iam.role => iam }
+  iam = concat(local.iam_role, local.iam_roles)
 
-  iam = module.iam
+  iam_map = { for iam in local.iam :
+    try(iam._key, "${iam.role}/${iam.condition._key}", "${iam.role}/${md5(jsonencode(iam.condition))}", iam.role) => iam
+  }
 }
 
 module "iam" {
@@ -18,7 +22,7 @@ module "iam" {
   module_enabled    = var.module_enabled
   module_depends_on = [var.module_depends_on]
 
-  folder = google_folder.folder[0].name
+  folder = try(google_folder.folder[0].name, null)
 
   role = each.key