+++ CHANGELOG.md
@@ -7,6 +7,23 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.2.0]
+
+### Added
+
+- Add test suite
+- Add support for multiple or no conditions on single IAM roles
+
+### Fixed
+
+- Fix references when `module_enabled = false`
+
+### Removed
+
+- BREAKING CHANGE: Remove support for Terraform before v1.0
+- BREAKING CHANGE: Remove support for Terraform Google Provider before v4.0
+- BREAKING CHANGE: Remove `module_enabled` output
+
 ## [0.1.1]
 
 ### Added
@@ -27,7 +44,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 <!-- markdown-link-check-disable -->
 
-[unreleased]: https://github.com/mineiros-io/terraform-google-folder/compare/v0.1.1...HEAD
+[unreleased]: https://github.com/mineiros-io/terraform-google-folder/compare/v0.2.0...HEAD
+[0.2.0]: https://github.com/mineiros-io/terraform-google-folder/compare/v0.1.1...v0.2.0
 [0.1.1]: https://github.com/mineiros-io/terraform-google-folder/compare/v0.1.0...v0.1.1
 [0.1.0]: https://github.com/mineiros-io/terraform-google-folder/compare/v0.0.1...v0.1.0
 [0.0.1]: https://github.com/mineiros-io/terraform-google-folder/releases/tag/v0.0.1